Fixed destroying objects by server
I have made some progress in regards to destrroying the objects - now for both banks we can drop the scroll without causing the network error - there is however another problem - when client is dropping the scroll we the score is increased by 2 instead of 1. From the host side it works as intended. I also noticed that I had a network error once - idk the reason, it didn't appear again but its worth to be careful in future and make sure it won't happen again.